(feature) Only show tasks for the currently signed in user
I've also added the current user's email address to the sign out link,
because it was getting confusing to know what tasks I should be able to
view at any given time.

The tests have been amended with `with(query: hash_including({}))`,
which permits any query parameters to be accepted by WebMock. This is
because otherwise we'd have lots of set up to get the user id to be
known *before* they sign in, to set up the mocks correctly.