(feature) Poll the API endpoint `GET /v1/submissions/:id` for the result of the levy calculation
(feature) Levy calculation is shown to the user so they can review it

The `ReviewsController` and `views/reviews` pages where used for polling both the ingest and
the calculate APIs. A `state` query params was added to the `review_processing_path` which will be
checked to identify the right url and method to call at the `ReviewsController#processing` action.

The validator class was also expanded to handle levy_completed status