RESTful #show for errored in_review submissions
Handling for a submission that needs reviewing and has errors on
entries.

Rather than use the Validator class I've chosen to add a new
Submission#errored_entries method as the errors page needs the entry as
well as the errors so that it can tell the user the sheet that the error
has been found on. This has the added advantage of removing the depency
on the Validator class entirely. It may be that with the more RESTful
approach, we don't need this class at all.

As with the previous #show-related commit, this takes the markup/content
from @meyric's #11-playback-screens branch.

Co-authored-by: mec <mec.rawlings@gmail.com>