(feature) Excel (.xls) files can be successfully uploaded
Previously, we had tests that appeared to be working, but were
misbehaving because Capybara's `attach_file` method ignores MIME types.

Because ActiveStorage was misdetecting Excel (.xls) files as
`application/octet-stream`, they were failing.

Since the detection isn't doing anything particularly clever, I think
it's good enough for now to change this look at check the file extension
only.

In the future, we should probably check the binary header of the file
or attempt to parse the file at the point of upload.

extension rather than mime