(feature) tools page
Add a tools page with instructions and a link to the URN list.

This is a temporary measure for August so that beta users can lookup URNs.
The file name includes the month as I would like users to be able to
identify the latest copy of the file at a glance, If it turns out we
need to update this file more than once a month we'll need to change
approach, but this is my happy place for now.

I've added a route, controller, view, feature spec and the file.