Wire in RESTful submissions#show to user flow
This gets us as far as viewing an "in_review" submission, assuming that
the levy calcution transitions the submission to that state (being done
separately on the API side).

The feature spec fakes the page auto-refresh by reloading the submission
page. Eventually we'll want a full-stack feature spec with JavaScript
enabled, but for now we are not set up to build specs that depend on
phantomjs.

Note that this work mostly makes the ReviewsController redundant. The
final step of actually completing a submission (and thus the task) is
yet to be done. That will follow after this.

The features that are removed here are now mostly redudant because they
are based on the old flow, where the levy calculation and other state
transitions were triggerd by the front-end app.