Remove N+1 API calls from tasks page
We can preload the task frameworks, rather than making an API call for
each and every task to look up its framework.

This also gives us a richer and more accurate stubbed response for
loading a list of tasks.

There is a slight issue in that it currently includes both the framework
and latest_submission for a task regardless of whether the API call
included it. The downside to this is that it means the stubbed response
doesn't necessarily match what would be returned in an actual API call,
which could result in false positives in the specs. For example, if the
`includes(:framework)` was removed from the tasks#index controller
action, the specs would still pass despite the fact that the page would
fail to render because of the missing frameworks. We need to look at a
better way to stub the API longer term, but for the sake of being ready
for the first public release, this is the compromise that makes sense.