Convert feature to request specs where appropriate
The feature specs that are simply testing a single endpoint are probably
more simply and appropriately tested as request specs, as they are
easier to write and require less setup.

The difference between a request spec and a feature spec according to
the [rspec-rails README][readme] is fairly subtle, but boils down to:
request specs are meant to test the request/response cycle, whilst
feature specs are for a richer end-to-end test that simulates browsing
the application itself. i.e. request specs are useful for testing single
endpoints, whilst feature specs are useful for simulating the user
behaviour with actual interactions on the page, and across multiple
requests.

This is partially to lay groundwork that will make it easier to add spec
coverage for the sign-in state across the various controllers.

[readme]: https://github.com/rspec/rspec-rails/blob/master/README.md