Add 'brakeman' to find potential vulnerabilities
This has been added to the default rake task that is run when you call
`bin/dspec` or `bin/drake` with no options. This will also be included
when CI gets run.

NB: I had to reorganise the rake tasks to ensure that RuboCop, RSpec and
Brakeman were run, as the original `rubocop.rake` was overwriting the
config.