Serve dynamic error pages
Serve the 404 and 500 pages via the app as a simple way to keep the
design and layout in sync between these pages and the app. These use a
simplified layout to minimise issues caused by session state, etc.

If we do end up needing to serve these as static pages later, we can use
these dynamic pages to generate them.