Adjust in-review valid submission screen
This is to make it more obvious how to re-upload a file if something
doesn't quite look right.

Louise requested that the 'Start again' button take the user back to the
task list, but right now, it's set up to always show you the most recent
submission. we'd need to put more thought and effort into reworking this
page so that the user could both see the existing submission file, and
upload a new file. Given that there is a separate piece of work to
rework the user journey, I've compromised here and had the button take
the user to the upload form instead.