Sync Submission file to API
This is the beginning of the work to move ingest and validation away
from lambdas and into background jobs performed on the API side. The
first step is to make sure the API knows where and how to access the
uploaded submission file.

At the moment the act of creating a submission by uploading a file
requires four separate API calls:

  1) create the submission record associated with the task
  2) create submission file record associated with the submission
  3) create the submission file blob associated with the submission file
  4) update the task to move it into an "in_progress" state

This could almost certainly be better encapsulated by putting it behind
a single endpoint that performs all of the above with a single POST
action, but that's an exercise for another time when we start looking at
rationalising the API interface into something more sensible.

I stopped short of pulling upload_file_submission out to its own object
because it seems that would make more sense at the point we have a
single endpoint for perform this operation.

This need to sync low-level file storage details using the API in this
way is a bit of a smell, but given the architecture as it has been
defined right now this feels like the most pragmatic way to make it so
that the supplier front-end can handle the file upload whilst still
having the backend API as the authoritative record of the file itself
without the backend having to read the file once it's been uploaded.

Note that the S3 upload file gets the submission file ID set as part of
its meta data. This is so that the ingest lambda can read it and use it
when creating entries, rather than creating a new submission file
against the submission, which is what it does currently. This does mean
that we'll need to co-ordinate the deploying of this change to coincide
with a corresponding change to the lambda to avoid a second submission
file being created and the entries being created against the wrong
submission file.