Add utility code and docs for adding new users
This adds documentation and supporting code for setting up new users on
the Report MI system so that they are able to sign in and make
submissions against their supplier's tasks. For GDPR reasons, a CSV for
October onboarding is not included here in the repository as it would
contain PPI for users.

The process is:

  - set up a user in Auth0
  - add a User record to the frontend database, including their Auth0
    ID, which makes it possible for the user to log in using Auth0.
  - Associate the newly created record with their Supplier in the API
    by creating a Membership record

When adding the user to Auth0 we set a random password and the user is
asked to use the "reset password" functionality in Auth0 to choose their
own the first time they attempt to log in.

We also set the supplier name and ID for each user in their Auth0 app
metadata so that they are available in the admin interface for Auth0 to
aid in support requests.