Show task description on tasks#index when present
This will be used to show the supplier name in the situation where a
user is associated with more than one supplier (and therefore has more
than one task on the same framework) so that they can differentiate
between tasks for their suppliers.

Note this is a temporary solution to the problem of users that need to
manage tasks for more than one supplier. We only discovered this was a
thing very close to the opening of the October submission window where
new frameworks and suppliers are being onboarded.