Update Tasks#index / Tasks#history to only request fields they need
Now the tasks endpoint on the API supports sparse fieldsets, we can
restrict which fields are returned, which cuts down the amount of
processing the API has to do.

We've updated the request mocks that hit this endpoint, so that any
future frontend changes which try to read other fields will result in
test failures.