User can download their submission files
On the `Task#show` page, make the submission file downloadable, by
clicking on its filename. This sends the user to the
`Submissions#download` action which then redirects the user to the
temporary (S3) download URL, that gets returned from the API.

The test has to be split in two, because Capybara's default driver does
not support remote URLs
(https://github.com/teamcapybara/capybara#racktest), which means
clicking the URL in the feature spec doesn't work. Instead, the redirect
to the download is carried out separately in a request spec.

NB: I updated the submission id used in `completed_task.json` to use the
`mocked_submission_id` value used elsewhere, in order to get the
download test to pass. For consistency, I also updated the id used in
`completed_task_with_no_business.json`.