Allow users to replace return in errors with no business
Once a user has started a business submission, they had no way of restarting the task and submitting no business instead. This led to support requests to reset tasks for users that wanted to do this.

This is resolved by simply putting this link into the new submission page, which is accessible from both the validation failed and in review states of an existsing submission.