Add a script to create user services
We use user services to provide credentials to our apps and environment
variables needed across all the apps. This script sources a file with
these in and then uses a json file to create all the ones we need.

The script also creates a routing for the admin interface so that only
some URLs will work.