Add a `.travis.yml`
As part of the move to the PaaS we will lose the infrastructure which
currently does CI and deployment. Therefore we need a replacement.
Because the repo is open we can use a free service such as travis.org or
CircleCI. CCS have a preference for travis so add a `.travis.yml`.

This configuration
- runs tests on ruby 2.5.5 because that is what we will run on the PaaS
- uses PostgreSQL 10 to match the PaaS