Allow downloading of Framework template from API
Now that the API can provide the S3 file key for the template file for a framework, we want to fetch that file and provide it to the supplier.

While all Frameworks might not have templates uploaded to them in the admin yet, we fallback to the existing templates included in this repository.