Control which branches are deployed to staging and production
When deploying to staging and production we want to stop people from
accidentally deploying the wrong branch. By default a cloudfoundry
deploy will push what ever files are present locally.

We deploy develop to staging and master to production.

For hot fixing we will allow people to force a deploy with `-f`