Fetch URN list from API and allow user to download it
The 'GET /v1/urn_lists' endpoint on the API responds with the following
attributes:

  - filename
  - file_key
  - byte_size

As with templates, the 'file_key' is used to allow the frontend to
stream the correct file stored in the S3 bucket. We also display the
size of the download, so this is included via 'byte_size'.