Store JWT when logging in, and include it in API requests
Previously, we were storing the `auth_id` returned from Auth0 and
passing that in the `X-Auth-Id` header on every API request. In order to
make the API public, we need to ensure that requests can only be made
by authorised users.

We do this by using the cryptographically-signed JWT which is returned
by Auth0 in the `id_token` attribute. This is stored in the `jwt`
attribute of the user's session.

When making requests to the mock API, we use a self-generated key pair
in order to sign the requests, so they verify as a real token would.