Fix occasional build failure
For some reason, on intermittent CI builds, AWS_S3_REGION is not set,
causing this error in the aws-partitions gem:

```
NoMethodError:
  undefined method `match' for nil:NilClass
```

Avoid using a real S3 client at all so we don't have to configure it by
supplying a spy instead, and check we talked to our fake S3 with the
file_key we mocked.

While here, change the title of the example, which referred to what it
used to do.