(feature) Handle the new management_charge_calculation_failed submission status
On the tasks page, use the same presentation as for a task whose
submission has failed validation.

However, given that this currently considered to be an edge case, which
can usually be resolved by submitting the return again, we decided that
it is sufficient to display a 500-like error page to the user when they
try to view the errors for the task, instead of new, informative error
message. (The full discussion around this can be found in Trello.)