(feature) Add Strict-Transport-Security header
This header would usually be added by Rails's force_ssl!, but for some
reason we have always had that commented-out. We are clearly taking care
of the HTTP -> HTTPS redirect at some other layer. So just add this
specific header here.

We must get permission from CCS to use the includeSubDomains directive
before we deploy this.