(fix) Fix Google Analytics by expanding Content-Security-Policy whitelist
With the Content-Security-Policy response header introduced in bc12b64,
I broke Google Analytics on production. It is able to load the script
from the whitelisted https://www.googletagmanager.com, but it turns out
that this script then attempts to load another script from
https://www.google-analytics.com.

This is the error that Chrome gives me:

gtm.js?id=GTM-TQNKLQS:361 Refused to load the script 'https://www.google-analytics.com/analytics.js' because it violates the following Content Security Policy directive: "script-src 'self' 'unsafe-inline' https://www.googletagmanager.com". Note that 'script-src-elem' was not explicitly set, so 'script-src' is used as a fallback.

I think I misread the (linked) documentation and assumed that you had to
_either_ whitelist the Google Tag Manager sources or the Google
Analytics sources. Looks like you might need to whitelist both.

Unfortunately we don't have any testing environment for Google Tag
Manager or Analytics, so will need to deploy this to prod again and see
if it fixes things.