Remove the simple_form gem
GitHub opened a security update PR for this gem, but it doesn't look
like we've actually ever made use of it in the codebase. There are some
initializers and library code, but the same code also appears in the
DataSubmissionServiceAPI repo. The gem _is_ used in that repo. I'm
guessing that we copied the code from there with the intention of using
it, but then never did so.