Show 'other_count' on the submission summary
Now we've exposed 'other_count' via the API, we can show it to the
user. We amended an existing summary scenario to add previously
missing row count expectations.

NB: The order of contracts and invoices have been swapped after
consulting with CCS. This is optimising for the common case, as we have
generally more invoices than contracts.