Test task display for a correction
When displaying a task the 'errors' flag is not shown and the status
message was not displayed.

Add coverage including a fixture that includes a task with a submission
in the status under test.

An existing feature needed updating to reflect that the task page now has
two task with correction.